separate component for skill dropdown menu
