+++ src/components/SkillDropdown/index.js
@@ -1,0 +1,2 @@
+import SkillDropdown from './SkillDropdown'
+export default SkillDropdown
